refactor: Move level/local/reader/* level/local/diskreader/*
