+++ internal/pkg/service/stream/storage/level/local/diskreader/reader.go
@@ -1,9 +1,9 @@
-// Package reader provides reading of tabular data from local storage for upload to staging storage.
+// Package diskreader provides reading of tabular data from local storage for upload to staging storage.
 // Data may be compressed on-tly-fly according to the configuration.
 // Regarding creating a reader, see:
 //   - The New function.
 //   - The "volume" package and the volume.OpenReader method in the package.
-package reader
+package diskreader
 
 import (
 	"context"
@@ -17,8 +17,8 @@ import (
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/compression"
 	compressionReader "github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/compression/reader"
 	compressionWriter "github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/compression/writer"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/level/local/diskreader/readchain"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/level/local/events"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/level/local/reader/readchain"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/model"
 	"github.com/keboola/keboola-as-code/internal/pkg/utils/errors"
 )
